chore(folder-structure): moves blog post components inside of components folder
