+++ src/components/blog/Details.tsx
