+++ src/components/blog/FigureImage.tsx
