+++ src/components/blog/StrikeThrough.tsx
