+++ src/components/blog/Strong.tsx
