+++ src/components/blog/index.ts
